Reboot galyleo project
Restarting active development of galyleo as part of a new effort to
rearchitect the collection of bash-based scripts currently used by
SDSC to manage HTTPS connections to Jupyter notebooks running on its HPC
systems through a reverse proxy service [1]. This initial version of the
new galyleo launch script focused on accessing the SDSC reverse proxy
service solves several on-going problems with the existing scripts:

  - completely command-line driven; attempts to avoid users editing
    scripts to make things work
  - Slurm batch job scripts are generated on the fly and placed in the
    jupyter --notebook_dir
  - supports the use of conda environments
  - supports the use of Singularity containers
  - supports the use of Environment Modules
  - allows Slurm GPU requests
  - provides a quiet option for formatted output, which will be required
    for future integration with the Open OnDemand deployment at SDSC
  - does not require the use of a jupyter config file; e.g.,
    JUPYTER_TOKEN is an environment variable never written out to a file
  - allows randomization of the JUPYTER_PORT as an obfuscation mechanism

Also included in this reboot commit is the 'tbs' directory, which holds
some of the work that has been in development on and off over the past
year or so with the aim of eventually updating the original, SSH
tunnel-based version of galyleo to a more production-ready state. This
SSH-based approach has been (and will likely continue to be) a low
priority effort though as the preferred method of launching Jupyter
notebooks on HPC Systems at SDSC moving forward will be through the
HTTPS reverse proxy service. However, the architecture of the galyleo
reboot will be designed to allow for the eventual inclusion of the
SSH-based approach to make this project more widely applicable for use
on non-SDSC systems.

[1] https://github.com/sdsc-hpc-training-org/reverse-proxy