Rename log.sh and its library of functions as slog
log is a command-line utility in BSD-based *nix distributions like Mac
OS X. If we want the new version of galyleo to allow remote Jupyter
notebook job requests over SSH as was done in the original, SSH
tunnel-based implementation, then any library with a 'log' function
within it will not work on Mac OS X because the shell will call the
existing 'log' utility instead. But a simple renaming of galyleo's
builtin logging library to 'slog' should avoid this issue.

[1] log(1)     BSD General Commands Manual     log(1)

NAME
     log -- Access system wide log messages created by os_log, os_trace and
     other logging systems.

SYNOPSIS
     log [command [options]]

     log help [command]
...