Updated slog.sh library; slog append
Restructured the slog.sh library such that all calls to the library are
now intended to go through the 'slog' function, which is the standard
interface used to control the execution of all other logging functions
in the slog library.

Another major change is that each individual logging function now
supports its own custom command-line interface, if necessary. At
present, however, most of the logging functions only expect a simple
-m (or --message) flag followed by the text associated with the log
message to be printed to the appropriate output location.

Added a new slog_append function that allows log messages to be appended
to a file. Note, however, this function was not created for logging
messages per se, but to provide a cleaner interface for building batch
job scripts with galyleo, which currently uses a simple echo and append
method in the 'launch' function.