Add call to reverse proxy service endpoint; linktoken
Added new call to the reverse proxy service to associate the batch job
id for the Jupyter notebook session to the connection token. This will
enable more detailed information for the user on the webpage displayed
to them while they are waiting for the job to start. Most importantly,
this will also allow us to provide dead-job notification to the user if
and when the job fails or gets cancelled for some reason.

Also simplified the redeemtoken call to the reverse proxy service.