Remove -M flag as short version of --mode command-line option
The -M flag was being used as the short version for both the --mode and
--memory-per-node command-line options in the galyleo launch command.
Removed -M from --mode to avoid a potential conflict. This change should
have no impact at this time since the --mode command-line option is not
currently in use, but simply being reserved for a future use case.