Add two calls to the destroytoken command
Added two calls to the destroytoken command from the Satellite reverse
proxy service's API, which is used to destroy the proxy mapping on
revocation of the issued token.

The first call to destroytoken revokes the token when the jupyter server
process is shutdown by the user. There remains the issue that this
reovacation will not occur if the scheduler kills the job. e.g., on a
TIMEOUT or OUT-OF-MEMORY error. One possible fix in the future, at least
for the TIMEOUT case, would be to run the jupyter server process using
the 'timeout' command-line utility to ensure there will always be some
interval of time at the end of a job to revoke the token.

The second call to destroytoken revokes the token when the batch job
submission to the scheduler fails for some reason.