Add input validation checks for Jupyter and Singularity executables
These validation checks pre-load the software environment defined by the
user before job submission and any interaction with the Satellite
reverse proxy service to see if the jupyter and/or singularity
executables are available. If they are not available within the software
environment, then the launch is halted and an error message is thrown to
inform the user of the problem.