Change job working directory to jupyter_notebook_dir
By default, the batch job generated and submitted to the scheduler by
galyleo will start with a present working directory (PWD) corresponding
to the job submission directory, which is the GALYLEO_CACHE_DIR where
the batch job scripts are stored. See also SLURM_SUBMIT_DIR and/or
PBS_O_WORDIR. This change to the generated batch job scripts actively
changes the PWD before jupyter is executed to be the
jupyter_notebook_dir set by the user. This will create a more consistent
environment for the user, matching their requested --notebook-dir
command-line argument for jupyter. Without this change a new shell
started from JupyterLab would start in the GALYLEO_CACHE_DIR.

Added the galyleo symlink to galyleo.sh. This should allow us to migrate
users away from using the *.sh extention. This will be useful in the
future, for example, if there is a rewrite of galyleo in Python (*.py),
which is something under consideration.

Also started updating README.md examples and usage guidance based on
recent user experiences and feedback.