Replace jupyter_notebook_dir cd access test with write permissions test
One issue that helped facilitate the problem outlined in the previous
commit was that every user in the training session had the same group
read and execute permissions on the instructors HOME directory, which
allowed each galyleo launch to pass the existing change directory test,
which only checked whether or not the user could access the
jupyter_notebook_dir, if it existed.

This new write permissions test is a more appropriate test for the
jupyter_notebook_dir than the change directory test AND it will help
protect against the problem outlined in the previous commit.