Allow dynamically created conda envs at runtime from an environment.yml file
This update adds the new --conda-yml command-line option, which allows
a user to specify an environment.yml that will be used to create a conda
environment at runtime prior to launching their Jupyter session. This
option requires a fast, node-local scratch disk available on the compute
resources requested for the session. If there is a standard scratch disk
location, this can now be set in the galyleo.conf file with the
GALYLEO_DEFAULT_LOCAL_SCRATCH_DIR environment variable. e.g., on Expanse [1].
Otherwise, the user must specify the --scratch-dir at runtime. The
default is /tmp.

[1] declare -xr GALYLEO_DEFAULT_LOCAL_SCRATCH_DIR='"/scratch/${USER}/job_${SLURM_JOB_ID}"'