Expanded the available system default variables in galyleo.conf
The galyleo configure command now also allows a system adminstrator to
create and deploy a galyleo.conf file with a default setting for the
number of cpus, memory, length of wallclock time limit, and a set of
environment modules to use when creating a Jupyter notebook session.
With the addition of these new default system variables, we hope to
simplify the launching of a minimal Jupyter notebook environment for
galyleo end users. For example, on Expanse, when using the standard
module environment's anaconda3 module as the default Jupyter environment,
a user would only need to provide their --account to start a Jupyter
notebook session with galyleo to get the default 1 CPU and 2 GB of
memory in the 'debug' partition for 30 minutes.