Add flag to bypass pre-launch checklist and control length of timeout
The new flag --disable-checklist allows the user to bypass most
of the validation checks/tests run prior to job submission to the
scheduler. This can be used when problems arise due erroneous
check/test failues on the (login) node where users launche Jupyter
notebook sessions.

One such problem observed previosuly on Expanse was that the time
required to check/test if the jupyter binary is available within a
Singularity container would become significant, hanging the launch
process when the load on the login node(s) was high. This is why the
timeout comand was prepended to this check/test. Here, we now also
add the --checklist-timeout option to adjust the length of the
timeout (in seconds), which has proven useful in the past as well.