Add option to --cache conda-pack'd environments built from yaml file
Removed existing --conda-pack command-line option for galyleo's launch
command. It was replaced with a new --cache command-line option to store
conda-pack'd  environments in the GALYLEO_CACHE_DIR. When enabled, any
conda environment built from an environment.yml file will be cached in
a named subdirectory of the GALYLEO_CACHE_DIR in a conda-pack'd tarball.
If a user attempts to launch another Jupyter notebook session with the
same environment.yml file, then the previosuly cached conda-pack'd
environment will be unpacked and reused instead of rebuilding the conda
environment from scratch again.