Change module purge to module reset
Originally, each galyleo launch command would begin the part of the
Slurm batch script where it constructs the job's software environment by
running Lmod's module purge command. However, the new behavior moving
forward will be to run the module reset command, which will unload all
of the modules a user currently has loaded, including the sticky ones,
and then it will reload the list of modules specified by
LMOD_SYSTEM_DEFAULT_MODULES. This should reduce the number of modules
that users may typically want to and/or need to reload in their launch
commands and allow these commands to be more version-agnostic in the
future.

https://github.com/mkandes/galyleo/issues/19