Add `GALYLEO_DEFAULT_MODULE`; fix abs and rel path issues with `--conda-yml`
Added the option to load a default software module in `galyleo.conf`. The
primary use case for this on SDSC systems is to load the `singularitypro`
module by default, which eliminates the need for users to specify it
explicitly in the `--env-modules` option to the `launch` command.

Fixed the previously poor handling of absolute and relativate paths when
using the `--conda-yml` option. The `environment.yml` no longer needs to
be located in the working directory where the `launch` command is run.

Also made a fairly extensive reorganization and simplification of the
README.md in order to highlight the use of newer, recommended features
in the `launch` command for newer users. All changes to `galyleo.sh`
should be backwards compatible. However, this new documentation reflects
the recent simplications in the `launch` command-line options and the
use of more implicit, default settings. e.g., GALYLEO_DEFAULT_MODULE.